Add files via upload
Java program that randomly generates words then asks the user to repeat the message. The time it takes for the user to respond is timed and later used to find the user's WPM (words per minute). The user can choose to keep using the program, each time generating a new set of words and retiming the response or they can choose to exit the program